renaming environment[s] folder
