+++ frontend/angular-fleets/src/app/services/group-listing-services/group-listing-fetch.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {map, Observable, tap} from "rxjs";
 import {GroupListingViewModel} from "../../models/group-listing/group-listing-view-model";
-import {environment} from '../../../environment/environment';
+import {environment} from '../../../environments/environment';
 
 @Injectable({
   providedIn: 'root'
