+++ frontend/angular-fleets/src/app/services/user-services/user.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {User} from "../../models/user/user";
 import {Observable} from "rxjs";
-import {environment} from '../../../environment/environment';
+import {environment} from '../../../environments/environment';
 
 @Injectable({
   providedIn: 'root'
